backend_api/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
pandas==2.2.3
openpyxl==3.1.2
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.9.0
